LPD-53392 Moving upgrade to object-rest-impl
