+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/upgrade/registry/ObjectRESTImplUpgradeStepRegistrator.java
@@ -1,0 +1,28 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.object.rest.internal.upgrade.registry;
+
+import com.liferay.object.rest.internal.upgrade.v1_0_0.SAPEntryServiceSignatureUpgradeProcess;
+import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Alejandro Tardín
+ */
+@Component(service = UpgradeStepRegistrator.class)
+public class ObjectRESTImplUpgradeStepRegistrator
+	implements UpgradeStepRegistrator {
+
+	@Override
+	public void register(Registry registry) {
+		registry.registerInitialization();
+
+		registry.register(
+			"0.0.1", "1.0.0", new SAPEntryServiceSignatureUpgradeProcess());
+	}
+
+}
