+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/upgrade/v1_0_0/SAPEntryServiceSignatureUpgradeProcess.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.security.service.access.policy.internal.upgrade.v3_0_2;
+package com.liferay.object.rest.internal.upgrade.v1_0_0;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
