+++ modules/apps/object/object-rest-test/build.gradle
@@ -29,7 +29,10 @@ dependencies {
 	testIntegrationImplementation project(":apps:portal-search:portal-search-api")
 	testIntegrationImplementation project(":apps:portal-security:portal-security-script-management-api")
 	testIntegrationImplementation project(":apps:portal-security:portal-security-script-management-test-util")
+	testIntegrationImplementation project(":apps:portal-security:portal-security-service-access-policy-api")
 	testIntegrationImplementation project(":apps:portal-vulcan:portal-vulcan-api")
+	testIntegrationImplementation project(":apps:portal:portal-upgrade-test-util")
 	testIntegrationImplementation project(":apps:sharing:sharing-api")
+	testIntegrationImplementation project(":apps:static:portal:portal-upgrade-api")
 	testIntegrationImplementation project(":test:arquillian-extension-junit-bridge")
 }
