+++ modules/apps/portal-security/portal-security-service-access-policy-service/src/main/java/com/liferay/portal/security/service/access/policy/internal/upgrade/registry/SAPServiceUpgradeStepRegistrator.java
@@ -9,7 +9,6 @@ import com.liferay.portal.kernel.upgrade.BaseSQLServerDatetimeUpgradeProcess;
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.security.service.access.policy.internal.upgrade.v3_0_0.util.SAPEntryTable;
 import com.liferay.portal.security.service.access.policy.internal.upgrade.v3_0_1.SAPEntryUpgradeProcess;
-import com.liferay.portal.security.service.access.policy.internal.upgrade.v3_0_2.SAPEntryServiceSignatureUpgradeProcess;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 
 import org.osgi.service.component.annotations.Component;
@@ -55,9 +54,6 @@ public class SAPServiceUpgradeStepRegistrator
 				new Class<?>[] {SAPEntryTable.class}));
 
 		registry.register("3.0.0", "3.0.1", new SAPEntryUpgradeProcess());
-
-		registry.register(
-			"3.0.1", "3.0.2", new SAPEntryServiceSignatureUpgradeProcess());
 	}
 
 }
